refactor: Move wallet subscription to app ready event
