+++ packages/neuron-wallet/src/main.ts
@@ -1,14 +1,17 @@
 import { app } from 'electron'
 import 'reflect-metadata'
+import { debounceTime } from 'rxjs/operators'
+
 import i18n from './utils/i18n'
 import { updateApplicationMenu } from './utils/application-menu'
-
 import Router from './router'
 import WindowManager from './models/window-manager'
 import createMainWindow from './startup/create-main-window'
 import createSyncBlockTask from './startup/sync-block-task/create'
 import initConnection from './database/address/ormconfig'
 import WalletsService from './services/wallets'
+import { WalletListSubject, CurrentWalletSubject } from './models/subjects/wallets'
+import dataUpdateSubject from './models/subjects/data-update'
 
 const walletsService = WalletsService.getInstance()
 
@@ -34,6 +37,20 @@ const openWindow = () => {
 }
 
 app.on('ready', async () => {
+  WalletListSubject.pipe(debounceTime(50)).subscribe(({ currentWallet = null, currentWalletList = [] }) => {
+    const walletList = currentWalletList.map(({ id, name }) => ({ id, name }))
+    const currentWalletId = currentWallet ? currentWallet.id : null
+    dataUpdateSubject.next({ dataType: 'wallet', actionType: 'update' })
+    updateApplicationMenu(walletList, currentWalletId)
+  })
+
+  CurrentWalletSubject.pipe(debounceTime(50)).subscribe(async ({ currentWallet = null, walletList = [] }) => {
+    updateApplicationMenu(walletList, currentWallet ? currentWallet.id : null)
+    if (currentWallet) {
+      dataUpdateSubject.next({ dataType: 'wallet', actionType: 'update' })
+    }
+  })
+
   i18n.changeLanguage(['zh', 'zh-CN'].includes(app.getLocale()) ? 'zh' : 'en')
   const wallets = walletsService.getAll()
   const currentWallet = walletsService.getCurrent()
